pyproject.toml
[project]
name = "04-meet-scheduler-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.6.2",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.1.1",
]
